+++ internal/user/models.go
@@ -3,7 +3,15 @@ package user
 import (
 	"time"
 
+	"github.com/google/uuid"
 	"gorm.io/gorm"
+)
+
+type RecentlyType string
+
+const (
+	RecentlyTypeUnspecified = "unspecified"
+	RecentlyTypeDao         = "dao"
 )
 
 type User struct {
@@ -12,4 +20,21 @@ type User struct {
 	UpdatedAt  time.Time
 	DeletedAt  gorm.DeletedAt `gorm:"index"`
 	DeviceUUID string
+}
+
+type RecentlyViewed struct {
+	gorm.Model
+
+	UserID uuid.UUID
+	Type   RecentlyType
+	TypeID string
+}
+
+func (rv *RecentlyViewed) TableName() string {
+	return "recently_viewed"
+}
+
+type RecentlyViewedList struct {
+	Views      []RecentlyViewed
+	TotalCount int64
 }