+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Implement recently viewed protocol
+
 ## [0.0.9] - 2023-07-27
 
 ### Added